requirements.txt
flask
tensorflow
torch
torchvision
numpy
opencv-python
matplotlib
pillow
scikit-learn
flask-cors
reportlab
